Cleanups. Removed unnecessary casts (400), null checks, etc.

+++ src/java/gnu/dtools/ritopt/LongOption.java
@@ -46,7 +46,7 @@ public class LongOption extends Option {
      */
 
     public LongOption() {
-	this( (long)0 );
+	this( 0 );
     }
 
     /**
